requirements.txt


Django>=2.0.12

celery>=3.1.25

coreschema>=0.0.4

django-appconf>=1.0.3

django-celery>=3.2.2

django-filter>=2.0.0

django-haystack>=2.8.1

django-pure-pagination>=0.3.0

django-ranged-response==0.2.0

django-shortuuidfield>=0.1.3

django-simple-captcha>=0.5.9

djangorestframework>=3.9.1

redis>=2.10.6